fix(jest): replace ts-jest config to transform
The following warning was generated when the test was run.

```
$ yarn workspace storycrawler test

yarn workspace v1.22.19
yarn run v1.22.19
$ npm run jest

ts-jest[ts-jest-transformer] (WARN) Define `ts-jest` config under `globals` is deprecated. Please do
transform: {
    <transform_regex>: ['ts-jest', { /* ts-jest config goes here in Jest */ }],
},
```

I have modified the settings as instructed.

Since the `diagnostics` option was set to false, I changed it to true as a behavior check and confirmed that a type error is displayed when the code is modified to cause a type error,
and that the type error is not displayed when the code is changed to false.